Improve native root store detection
